feat: ✨ move configs to config folder and adjust refs
