+++ package.json
@@ -39,7 +39,7 @@
     "@ensdomains/address-encoder": "0.2.6",
     "@ensdomains/ens-contracts": "0.0.5",
     "@myetherwallet/eth2-keystore": "0.2.3",
-    "@myetherwallet/mew-components": "0.5.89-beta",
+    "@myetherwallet/mew-components": "0.5.90-beta",
     "@myetherwallet/mewconnect-web-client": "2.1.23-beta.9",
     "@sentry/browser": "6.8.0",
     "@sentry/tracing": "6.8.0",
