+++ src/modules/access-wallet/ModuleAccessWalletHardware.vue
@@ -229,7 +229,7 @@ import { mapActions, mapGetters, mapState } from 'vuex';
 import WALLET_TYPES from '@/modules/access-wallet/common/walletTypes';
 import Web3 from 'web3';
 import { fromWei } from 'web3-utils';
-import { ROUTES_WALLET } from '@/core/router/config-routes';
+import { ROUTES_WALLET } from '@/core/configs/configRoutes';
 const MAX_ADDRESSES = 5;
 
 export default {
