+++ src/modules/access-wallet/ModuleAccessWalletSoftware.vue
@@ -88,7 +88,7 @@ import { mapActions, mapState } from 'vuex';
 import { Toast, ERROR } from '@/modules/toast/handler/handlerToast';
 import { SOFTWARE_WALLET_TYPES } from './software/handlers/helpers';
 import handlerAccessWalletSoftware from './software/handlers/handlerAccessWalletSoftware';
-import { ROUTES_WALLET } from '../../core/router/config-routes';
+import { ROUTES_WALLET } from '../../core/configs/configRoutes';
 
 export default {
   name: 'ModuleAccessWalletSoftware',
