+++ src/modules/create-wallet/components/CreateWalletMnemonicPhrase.vue
@@ -236,7 +236,7 @@
 import mnemonicPhraseTable from '@/components/MnemonicPhraseTable';
 import phraseBlock from '@/components/PhraseBlock';
 import { Toast, ERROR } from '@/modules/toast/handler/handlerToast';
-import { ROUTES_HOME } from '@/core/router/config-routes';
+import { ROUTES_HOME } from '@/core/configs/configRoutes';
 export default {
   name: 'CreateWalletMnemonicPhrase',
   components: {
