+++ src/views/components-default/HomeLanding.vue
@@ -100,7 +100,7 @@
 </template>
 
 <script>
-import { ROUTES_HOME } from '@/core/router/config-routes';
+import { ROUTES_HOME } from '@/core/configs/configRoutes';
 
 export default {
   name: 'HomeLanding',
