+++ src/views/components-default/TheDefaultFooter.vue
@@ -300,7 +300,7 @@
 </template>
 
 <script>
-import { ROUTES_HOME } from '@/core/router/config-routes';
+import { ROUTES_HOME } from '@/core/configs/configRoutes';
 
 export default {
   name: 'TheDefaultFooter',
