+++ src/views/layouts-default/TheToolsLayout.vue
@@ -51,7 +51,7 @@ import ModuleToolsWatchOnly from '@/modules/tools/ModuleToolsWatchOnly';
 import ModuleToolsConvert from '@/modules/tools/ModuleToolsConvert';
 import ModuleToolsOfflineHelper from '@/modules/tools/ModuleToolsOfflineHelper';
 import ModuleMessageVerify from '@/modules/message/ModuleMessageVerify';
-import { ROUTES_HOME } from '@/core/router/config-routes';
+import { ROUTES_HOME } from '@/core/configs/configRoutes';
 
 export default {
   name: 'MoreActions',
